requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.3
fonttools==4.42.1
graphviz==0.20.1
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
only-train-once==2.0.16
onnx==1.15.0
packaging==23.1
Pillow==10.0.0
protobuf==4.25.1
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
scipy==1.11.3
six==1.16.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.16.2
